+++ examples/dark-space/next.config.mjs
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   output: "export",
-  // basePath: "/examples/navbar-153",
+  // basePath: "/examples/hanav",
 };
 
 export default nextConfig;