+++ examples/demo-nextjs/components/docs/index.jsx
@@ -6,8 +6,8 @@ export default async function Docs({ lng }) {
   const { t } = await useTranslation(lng);
   const desc = t("desc");
   return <>
-    <h1 className={styles.txt}>navbar-153</h1>
-    <pre className={styles.install}>npm install navbar-153</pre>
+    <h1 className={styles.txt}>hanav</h1>
+    <pre className={styles.install}>npm install hanav</pre>
     <p className={styles.txt}>{desc}</p>
     <div className={styles.feats}>
       <FeatSlate emoji="🍯" title={t("smooth_t")} desc={t("smooth_d")} />
@@ -15,10 +15,10 @@ export default async function Docs({ lng }) {
       <FeatSlate emoji="♿️" title={t("a11y_t")} desc={t("a11y_d")} />
       <FeatSlate emoji="🎨" title={t("style_t")} desc={t("style_d")} />
     </div>
-    <p className={styles.txt}><Trans i18nKey="usage" t={t}>0<a href="https://github.com/wswmsword/navbar-153">1</a>2</Trans></p>
+    <p className={styles.txt}><Trans i18nKey="usage" t={t}>0<a href="https://github.com/wswmsword/hanav">1</a>2</Trans></p>
     <p className={styles.txt}>{t("stars")}</p>
-    <div aria-hidden="true" className={`${styles.presentation} ${styles.textShadow}`}>NAVBAR-153</div>
-    <div aria-hidden="true" className={styles.presentation2}>NAVBAR-153</div>
+    <div aria-hidden="true" className={`${styles.presentation} ${styles.textShadow}`}>HANAV</div>
+    <div aria-hidden="true" className={styles.presentation2}>HANAV</div>
   </>;
 }