+++ examples/demo/src/components/navbar-slate/index.jsx
@@ -1,10 +1,9 @@
 import styles from "./index.module.css";
 
 export default function NavbarSlate({ propsFromN, head, tail, dynamicWidth }) {
-  const contentItemStyle = {
-    width: dynamicWidth ? 300 : "100%",
-    flexShrink: 0,
-  };
+  const contentItemStyle = dynamicWidth ? {
+    width: 300,
+  } : {};
   return <div
     className={styles.wrapper}
     {...propsFromN}
@@ -17,10 +16,10 @@ export default function NavbarSlate({ propsFromN, head, tail, dynamicWidth }) {
       <FeatSlate e="🎨" t="高度自定义" d="为每个元素设置样式，让每个导航栏变得不一样" />
     </div>
     <ul className={styles.links}>
-      <li><a ref={head} href="https://github.com/wswmsword/navbar-153?tab=readme-ov-file#%E5%AE%89%E8%A3%85%E5%92%8C%E4%BD%BF%E7%94%A8">安装和使用</a></li>
-      <li><a href="https://github.com/wswmsword/navbar-153?tab=readme-ov-file#api">API</a></li>
-      <li><a href="https://github.com/wswmsword/navbar-153?tab=readme-ov-file#%E9%94%AE%E7%9B%98%E4%BA%A4%E4%BA%92">键盘交互</a></li>
-      <li><a ref={tail} href="https://github.com/wswmsword/navbar-153?tab=readme-ov-file#%E5%BC%80%E5%8F%91%E4%B8%8E%E7%BB%B4%E6%8A%A4%E6%96%B9%E5%90%91">开发与维护方向</a></li>
+      <li><a ref={head} href="https://github.com/wswmsword/hanav?tab=readme-ov-file#%E5%AE%89%E8%A3%85%E5%92%8C%E4%BD%BF%E7%94%A8">安装和使用</a></li>
+      <li><a href="https://github.com/wswmsword/hanav?tab=readme-ov-file#api">API</a></li>
+      <li><a href="https://github.com/wswmsword/hanav?tab=readme-ov-file#%E9%94%AE%E7%9B%98%E4%BA%A4%E4%BA%92">键盘交互</a></li>
+      <li><a ref={tail} href="https://github.com/wswmsword/hanav?tab=readme-ov-file#%E5%BC%80%E5%8F%91%E4%B8%8E%E7%BB%B4%E6%8A%A4%E6%96%B9%E5%90%91">开发与维护方向</a></li>
     </ul>
   </div>;
 }