+++ examples/demo/src/main.jsx
@@ -5,13 +5,25 @@ import Header from "./components/header";
 import Footer from "./components/footer";
 import CenterBox from "./components/center-box";
 import "./index.css"
+import { createContext } from "react";
+import { useState } from "react";
 
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <>
-    <Header />
-    <CenterBox>
-      <Docs />
-      <Footer />
-    </CenterBox>
-  </>
-)
+export const Context = createContext();
+export const ContextF = createContext();
+
+const container = ReactDOM.createRoot(document.getElementById("root"));
+
+container.render(<App />)
+
+function App() {
+  const [go, setG] = useState(false);
+  return <Context.Provider value={{ go }}>
+    <ContextF.Provider value={setG}>
+      <Header />
+      <CenterBox>
+        <Docs />
+        <Footer />
+      </CenterBox>
+    </ContextF.Provider>
+  </Context.Provider>
+}